refactor: 불필요한 view 폴더 제거 및 import 경로 수정
